first pass at making modular enough to be a CLI tool and called from other packages
