+++ .gitignore
@@ -1,0 +1,4 @@
+*~
+pkg
+src
+cmd/slackcat
